+++ CHANGELOG.rst
@@ -17,6 +17,7 @@ Changes:
 ^^^^^^^^
 
 - Added ``environ.generate_help`` to the public interface. It is implemented by...
+
   * ``environ._environ_config.generate_help``
   * ``environ._environ_config._generate_help_dicts``
   * ``environ._environ_config._format_help_dicts``