+++ CHANGELOG.rst
@@ -23,8 +23,9 @@ Deprecations:
 Changes:
 ^^^^^^^^
 
-- Added ``environ.generate_help(AppConfig)`` to create a help string based on the configuration.
-- Added ``AppConfig.from_environ()`` to instantiate the config class without importing the ``environ`` module.
+- Added ``environ.generate_help(AppConfig)`` and ``AppConfig.generate_help()`` (two alternatives) to create a help string based on the configuration.
+- Added ``AppConfig.from_environ()`` to instantiate the config class. This is alternative to ``environ.config()``.
+- ``environ.config(from_environ="fr_env", generate_help="gen_hlp")`` allows passing alternative names for class methods "from_environ" and ``"generate_help"``, or prevent their creation by pasing ``None``.
 - If ``environ.var`` is passed an ``attr.Factory``, the callable is used to generate the default value.